+++ trunk/modules/cloudsim/src/test/java/org/cloudbus/cloudsim/lists/PeListTest.java
@@ -1,3 +1,11 @@
+/*
+ * Title:        CloudSim Toolkit
+ * Description:  CloudSim (Cloud Simulation) Toolkit for Modeling and Simulation of Clouds
+ * Licence:      GPL - http://www.gnu.org/copyleft/gpl.html
+ *
+ * Copyright (c) 2009-2010, The University of Melbourne, Australia
+ */
+
 package org.cloudbus.cloudsim.lists;
 
 import static org.junit.Assert.assertEquals;
@@ -15,8 +23,12 @@ import org.cloudbus.cloudsim.provisioners.PeProvisionerSimple;
 import org.junit.Before;
 import org.junit.Test;
 
+/**
+ * @author		Anton Beloglazov
+ * @since		CloudSim Toolkit 2.0
+ */
 public class PeListTest {
-	
+
 	private static final double MIPS = 1000;
 
 	private List<Pe> peList;
@@ -24,23 +36,23 @@ public class PeListTest {
 	@Before
 	public void setUp() throws Exception {
 		peList = new ArrayList<Pe>();
-		
+
 		peList.add(new Pe(0, new PeProvisionerSimple(MIPS)));
 		peList.add(new Pe(1, new PeProvisionerSimple(MIPS)));
 	}
-	
+
 	@Test
 	public void testGetMips() {
 		assertEquals(MIPS, PeList.getMips(peList, 0), 0);
 		assertEquals(MIPS, PeList.getMips(peList, 1), 0);
 		assertEquals(-1, PeList.getMips(peList, 2));
 	}
-	
+
 	@Test
 	public void testGetTotalMips() {
 		assertEquals(MIPS * peList.size(), PeList.getTotalMips(peList), 0);
 	}
-	
+
 	@Test
 	public void testSetPeStatus() {
 		assertEquals(2, PeList.getFreePesNumber(peList));
@@ -57,7 +69,7 @@ public class PeListTest {
 		assertEquals(0, PeList.getFreePesNumber(peList));
 		assertEquals(2, PeList.getBusyPesNumber(peList));
 	}
-	
+
 	@Test
 	public void testSetStatusFailed() {
 		assertEquals(Pe.FREE, PeList.getById(peList, 0).getStatus());
@@ -68,7 +80,7 @@ public class PeListTest {
 		PeList.setStatusFailed(peList, false);
 		assertEquals(Pe.FREE, PeList.getById(peList, 0).getStatus());
 		assertEquals(Pe.FREE, PeList.getById(peList, 1).getStatus());
-		
+
 		PeList.setStatusFailed(peList, "test", 0, true);
 		assertEquals(Pe.FAILED, PeList.getById(peList, 0).getStatus());
 		assertEquals(Pe.FAILED, PeList.getById(peList, 1).getStatus());
@@ -76,7 +88,7 @@ public class PeListTest {
 		assertEquals(Pe.FREE, PeList.getById(peList, 0).getStatus());
 		assertEquals(Pe.FREE, PeList.getById(peList, 1).getStatus());
 	}
-	
+
 	@Test
 	public void testFreePe() {
 		assertSame(peList.get(0), PeList.getFreePe(peList));
@@ -85,28 +97,28 @@ public class PeListTest {
 		PeList.setPeStatus(peList, 1, Pe.BUSY);
 		assertNull(PeList.getFreePe(peList));
 	}
-	
+
 	@Test
 	public void testGetMaxUtilization() {
 		Vm vm0 = new Vm(0, 0, MIPS / 2, 1, 0, 0, 0, 0, "", null);
 		Vm vm1 = new Vm(1, 0, MIPS / 2, 1, 0, 0, 0, 0, "", null);
-		
-		assertTrue(peList.get(0).getPeProvisioner().allocateMipsForVm(vm0, MIPS / 3)); 
+
+		assertTrue(peList.get(0).getPeProvisioner().allocateMipsForVm(vm0, MIPS / 3));
 		assertTrue(peList.get(1).getPeProvisioner().allocateMipsForVm(vm1, MIPS / 5));
-		
-		assertEquals((double) (MIPS / 3) / MIPS, PeList.getMaxUtilization(peList), 0.001);
+
+		assertEquals((MIPS / 3) / MIPS, PeList.getMaxUtilization(peList), 0.001);
 	}
 
 	@Test
 	public void testGetMaxUtilizationAmongVmsPes() {
 		Vm vm0 = new Vm(0, 0, MIPS / 2, 1, 0, 0, 0, 0, "", null);
 		Vm vm1 = new Vm(1, 0, MIPS / 2, 1, 0, 0, 0, 0, "", null);
-		
-		assertTrue(peList.get(0).getPeProvisioner().allocateMipsForVm(vm0, MIPS / 3)); 
+
+		assertTrue(peList.get(0).getPeProvisioner().allocateMipsForVm(vm0, MIPS / 3));
 		assertTrue(peList.get(1).getPeProvisioner().allocateMipsForVm(vm1, MIPS / 5));
-			
-		assertEquals((double) (MIPS / 3) / MIPS, PeList.getMaxUtilizationAmongVmsPes(peList, vm0), 0.001);
-		assertEquals((double) (MIPS / 5) / MIPS, PeList.getMaxUtilizationAmongVmsPes(peList, vm1), 0.001);
+
+		assertEquals((MIPS / 3) / MIPS, PeList.getMaxUtilizationAmongVmsPes(peList, vm0), 0.001);
+		assertEquals((MIPS / 5) / MIPS, PeList.getMaxUtilizationAmongVmsPes(peList, vm1), 0.001);
 	}
 
 }
